+++ src/Randomizer/index.js
@@ -81,6 +81,7 @@ export async function execute(romfile, specs, hooks) {
   Dialog.afterOrder = [PsiTeleportObject];
   EnemyPlaceObject.afterOrder = [BattleEntryObject];
   MapEnemyObject.afterOrder = [AncientCave];
+  MapMusicObject.afterOrder = [AncientCave];
   MapSpriteObject.afterOrder = [PsiTeleportObject, AncientCave];
 
   try {